Merged PR 3953: Updated CI to test all supported version
Changes made:
BUILD: Updated CI to test all supported versions. Also, check to make sure the correct version of PHP is installed and used.
BUG: Constant modifier is not supported until 7.1.0 so remove from MissingPropertyReason.php
BUILD: Added a composer-dev.json to reference a local package in development CI. This is currently used by device-detection-php-onpremise and is reference a pipeline-php-core that resides in the same folder. Moving forwards for CI, we might want to add the pipeline-php-core as a submodule and reference that instead. This way it might impact device-detection-php-onpremise but will support development build and test without waiting for pipeline-php-core to be released.